// src/generated/segmentation.ts
export const candidateSurfaces = [
  {
    "surface": "cyto",
    "partId": "combining:cyt-o",
    "transformations": []
  },
  {
    "surface": "cyt",
    "partId": "combining:cyt-o",
    "transformations": [
      {
        "kind": "drop_terminal_vowel",
        "vowel": "o"
      }
    ]
  },
  {
    "surface": "glyco",
    "partId": "combining:glyc-o",
    "transformations": []
  },
  {
    "surface": "glyc",
    "partId": "combining:glyc-o",
    "transformations": [
      {
        "kind": "drop_terminal_vowel",
        "vowel": "o"
      }
    ]
  },
  {
    "surface": "ad",
    "partId": "prefix:ad",
    "transformations": []
  },
  {
    "surface": "brady",
    "partId": "prefix:brady",
    "transformations": []
  },
  {
    "surface": "endo",
    "partId": "prefix:endo",
    "transformations": []
  },
  {
    "surface": "hyper",
    "partId": "prefix:hyper",
    "transformations": []
  },
  {
    "surface": "hypo",
    "partId": "prefix:hypo",
    "transformations": []
  },
  {
    "surface": "peri",
    "partId": "prefix:peri",
    "transformations": []
  },
  {
    "surface": "tachy",
    "partId": "prefix:tachy",
    "transformations": []
  },
  {
    "surface": "aden",
    "partId": "root:aden",
    "transformations": []
  },
  {
    "surface": "adeno",
    "partId": "root:adeno",
    "transformations": []
  },
  {
    "surface": "adren",
    "partId": "root:adren",
    "transformations": []
  },
  {
    "surface": "adreno",
    "partId": "root:adreno",
    "transformations": []
  },
  {
    "surface": "arthr",
    "partId": "root:arthr",
    "transformations": []
  },
  {
    "surface": "arthro",
    "partId": "root:arthro",
    "transformations": []
  },
  {
    "surface": "card",
    "partId": "root:card",
    "transformations": []
  },
  {
    "surface": "cardi",
    "partId": "root:cardi",
    "transformations": []
  },
  {
    "surface": "cardio",
    "partId": "root:cardio",
    "transformations": []
  },
  {
    "surface": "derm",
    "partId": "root:derm",
    "transformations": []
  },
  {
    "surface": "dermat",
    "partId": "root:dermat",
    "transformations": []
  },
  {
    "surface": "dermato",
    "partId": "root:dermato",
    "transformations": []
  },
  {
    "surface": "gastr",
    "partId": "root:gastr",
    "transformations": []
  },
  {
    "surface": "gastro",
    "partId": "root:gastro",
    "transformations": []
  },
  {
    "surface": "hepat",
    "partId": "root:hepat",
    "transformations": []
  },
  {
    "surface": "hepato",
    "partId": "root:hepato",
    "transformations": []
  },
  {
    "surface": "nephr",
    "partId": "root:nephr",
    "transformations": []
  },
  {
    "surface": "nephro",
    "partId": "root:nephro",
    "transformations": []
  },
  {
    "surface": "neur",
    "partId": "root:neur",
    "transformations": []
  },
  {
    "surface": "neuro",
    "partId": "root:neuro",
    "transformations": []
  },
  {
    "surface": "oste",
    "partId": "root:oste",
    "transformations": []
  },
  {
    "surface": "osteo",
    "partId": "root:osteo",
    "transformations": []
  },
  {
    "surface": "ren",
    "partId": "root:ren",
    "transformations": []
  },
  {
    "surface": "reno",
    "partId": "root:reno",
    "transformations": []
  },
  {
    "surface": "thromb",
    "partId": "root:thromb",
    "transformations": []
  },
  {
    "surface": "thrombo",
    "partId": "root:thrombo",
    "transformations": []
  },
  {
    "surface": "al",
    "partId": "suffix:al",
    "transformations": []
  },
  {
    "surface": "algia",
    "partId": "suffix:algia",
    "transformations": []
  },
  {
    "surface": "cyte",
    "partId": "suffix:cyte",
    "transformations": []
  },
  {
    "surface": "emia",
    "partId": "suffix:emia",
    "transformations": []
  },
  {
    "surface": "ia",
    "partId": "suffix:ia",
    "transformations": []
  },
  {
    "surface": "itis",
    "partId": "suffix:itis",
    "transformations": []
  },
  {
    "surface": "kine",
    "partId": "suffix:kine",
    "transformations": []
  },
  {
    "surface": "megaly",
    "partId": "suffix:megaly",
    "transformations": []
  },
  {
    "surface": "oma",
    "partId": "suffix:oma",
    "transformations": []
  },
  {
    "surface": "opathy",
    "partId": "suffix:opathy",
    "transformations": []
  },
  {
    "surface": "plasty",
    "partId": "suffix:plasty",
    "transformations": []
  },
  {
    "surface": "scopy",
    "partId": "suffix:scopy",
    "transformations": []
  }
] as const;